refactor(auth): refactor routes
