+++ custom/emails/activate-account/html.ejs
@@ -22,7 +22,7 @@
   <body>
     <h2>Hello <%= display_name %></h2>
     <p>Please confirm your email address by clicking the button below:</p>
-    <a href="<%= url %>/auth/account/activate?ticket=<%= ticket %>">
+    <a href="<%= url %>/auth/activate?ticket=<%= ticket %>">
       <button>Confirm email address</button></a
     >
     <p>Thanks,<br />The Team</p>
