+++ src/routes/auth/activate.ts
