+++ src/routes/auth/delete.test.ts
@@ -5,7 +5,7 @@ import { SuperTest, Test, agent } from 'supertest'
 import { PORT } from '@shared/config'
 import { registerAccount } from '@test/test-utils'
 
-import { app } from '../../../server'
+import { app } from '../../server'
 
 let request: SuperTest<Test>
 
@@ -24,6 +24,6 @@ afterAll(async () => {
 
 it('should delete an account', async () => {
   await registerAccount(request)
-  const { status } = await request.post('/auth/account/delete')
+  const { status } = await request.post('/auth/delete')
   expect(status).toEqual(204)
 })
