+++ src/routes/auth/delete.ts
